Implemented some exemples of subtraction in Main.java and changed integers into double

+++ src/Operations/Subtraction.java
@@ -3,13 +3,13 @@ package Operations;
 public class Subtraction {
 
     //Metodo sottrazione a più elementi
-    public static int sub(int a, int b) {
+    public static double sub(double a, double b) {
         return a - b;
     }
 
     //Metodo sottrazione degli elementi di un array
-    public static int sub(int[] values) {
-        int result = values[0];
+    public static double sub(double[] values) {
+        double result = values[0];
         for (int i = 1; i < values.length; i++) {
             result -= values[i];
         }
